Automated g4 rollback of commit afe4bc340340809260a82d8efc4d5654ac42ad88.
*** Reason for rollback ***

Rollforward with fix

To have stdout/stderr in the BuildEventStream, the BuildEventStreamerModule
registers a listener that accumulates the values for the streamer to
collect it and regularly send it over the stream. As we have to also catch
stdout/stderr that happens early in the build, we have to register the listener
before(!) the options are parsed; therefore it is registered unconditionally.
Now, if there is no streamer created after parsing the options there is no
one to collect the data and it grows indefinitely. Fix this, by disabling
the collection in this case.

*** Original change description ***

Automated g4 rollback of commit e051640f5f3738f1e95ca6e92e8dca9ec8910dd4.

*** Reason for rollback ***

Results in NegativeArraySizeExceptions when there's a high volume of data.

*** Original change description ***

BEP: Report stdout/stderr

By recording registering a properly synchronized OutErr as listener
and providing it as OutErrProvider to the BuildEventStreamer.

Change-Id: Id553fcdb85327be28561634268511304fcc2ad3f
PiperOrigin-RevId: 155374710