Add a custom interface for cache hit processing in actions
The new interface mirrors ActionExecutionContext, but is restricted to exactly
the parts used right now. I did consider using ActionExecutionContext, but it
contains some parts that we don't want to make available for cache hits.

The end goal is to allow the build event stream access to artifact metadata,
in particular for TestResult and TestSummary events, which in turn requires
making artifact metadata available when the TestRunnerAction is a cache hit.

PiperOrigin-RevId: 155492447