Fix aspect processing when same dep appears in 2 attributes.
If a target appeared in 2 different attributes, it is not processed twice, even if different aspects were applied to the different attributes. In that case, only one of the aspects is applied. This commit fixes this by checking which aspects have been applied to the target, instead of checking if the target was already processed.

PiperOrigin-RevId: 156738275