Add a #hashCode() method to WorkspaceStatusAction implementations as is advisable when overriding #equals().
Intricacies within Skyframe prevented us from adding this method before, but whatever those were are not a problem anymore. This was empirically proven by making the change, looking at the set of tests that failed when I last tried this, then finding that they don't fail.

The reason for the original failures will stay obscured by the mists of time forever.

RELNOTES: None.
PiperOrigin-RevId: 156835875