Improve error messages for malformed Android SDK/NDK.
Currently, if the user has an SDK/NDK that is missing required directories, we
print out an InconsistentFileSystemException. This CL will give a better error,
indicating which directory is missing (or is not a directory).

Implementation is roughly taken from NewLocalRepositoryFunction.

Followup to https://github.com/bazelbuild/bazel/issues/2739.

RELNOTES: None
PiperOrigin-RevId: 156913532