process-tools: Fail silently when we can't install a signal handler.
linux-sandbox expects InstallSignalHandler to fail silently when called
for a signal that doesn't allow one to install a signal handler. This
behavior was accidentally lost and changed into a DIE() on sigaction
failure when unifying the signal handling code with process-wrapper in
https://github.com/bazelbuild/bazel/commit/487233a4860ef6e9ed1ba03abdce61589d9e1016. This CL restores the previous behavior.

The issue results in this failure:

$ linux-sandbox -- /bin/true
third_party/bazel/src/main/tools/process-tools.cc:112: "sigaction": Invalid argument

We do have tests for linux-sandbox and Bazel's usage of the linux-sandbox
strategy. None of these actually started failing when the linux-sandbox
stopped working, due to our automatic fallback logic that we added in
order to not annoy users / CI on platforms that don't support the Linux
sandbox :( Bazel silently falls back to the processwrapper-sandbox
strategy and disables the entire linux-sandbox test suite, but signals
to CI that it passed (because we don't support a "Skipped" test status).
Well, congrats. What an epic fail. I will have to rework this next week.
PiperOrigin-RevId: 157021455