Remote worker: skip non-existent files after action execution
Tests basically always finish with non-existent files, and this is not an
error on the server side. Instead, the client (Bazel) checks whether all the
files it expects are present.

The test didn't catch this because it was silently falling back to local
execution.

Non-existent files were leading to an IOException, which caused the remote
worker to log nothing, and silently return an error with no output. Log
errors in the remote worker to make future debugging easier.

Fixes #2887.

PiperOrigin-RevId: 157400131