BEP: Correctly report test exit code.
When a test failure occurred the BEP would still report a build_finished event
with exit code SUCCESS, even though bazel's exit code was TESTS_FAILED. This is
because we would rely on the exit code reported by the BuildCompleteEvent. However,
this exit code contains only the build status without taking into account the test status.

In this change we introduce the TestingCompleteEvent that reports bazel's final
exit code in case of "bazel test".

RELNOTES: None.
PiperOrigin-RevId: 157445808