When filtering resources in analysis, also filter resource roots
If all resources from a directory are filtered out, the resource processing
action should not receive a reference to that directory.

Filtering the entire contents of a directory is uncommon but not impossible. It
may be indicative of an error (if the user filters on a language, but only
provided resources in other languages - in this case, the build should and will
fail during Java compilation when the requested resource does not exist) or of
weird but acceptable behavior (placing resources for some default language in
one resource directory and resources for all other languages in some other
directory, then inheriting both directories and filtering on language) that
should be correctly handled.

RELNOTES: none
PiperOrigin-RevId: 157499745