Windows: Create a dangling junction when symlink target doesn't exist
On Unix system, `ln -s foo bar` will create an dangling symlink bar -> foo
even foo doesn't exist. The current implementation of createSymbolicLink for
Windows fails in this situtation. And this lead to #2474.

In this change, we create a dangling junction when the target doesn't exist
which mimics the behavior on Unix.

Fixed https://github.com/bazelbuild/bazel/issues/2474

Change-Id: I442ca3e2fb20b76c9b5bbfee903299fe51481f43
PiperOrigin-RevId: 157694631