Experimental UI: don't show dates for no-build commands
When building with timestamps enabled, it is useful to know the date
as well, e.g., when later looking at logs. This, however, is not the
case if the command does not build (e.g., "bazel help", "bazel query"),
or, in general, if the first output is only produced after the command
is completed.

Change-Id: I75ef38fbb98e886b1dc38899efa10188055f87e2
PiperOrigin-RevId: 157700578