Make compatible_with = ["all", "foo"] the same as compatible_with = ["all"].
Assuming "all" fulfills "foo", these should be exactly the same (and
maybe we should trigger a redundant listing error). In practice, it's
possible to make the first case succeed while the second fails because of
environment refining and lack of static constraint checking for selects.

See changes for details.

Also refactor ConstraintSemantics.checkConstraints to divide and conquer more clearly.

PiperOrigin-RevId: 158047217