experimental_ui, rate limiting test: work around output by the  client
When the experimental UI is told to rate limit the amount of output
generated, this limit does not include startup messages generated by
the client. Make the test aware of that behaviour by allowing a bit
more output than the UI limit, to take into account the client startup
messages. To keep the test still strict increase the output generated
by the actions a lot, so that the client output is small compared to
what we expect and limit.

PiperOrigin-RevId: 158386698