Automated g4 rollback of commit c4ece4f53bc0c836aae3d4e9adf54f8b1cadee36.
NO_SQ=Only failing TAP targets in our project are from an unrelated bug.

*** Reason for rollback ***

Initially we thought this broke just a few builds that were fixed in the depot. The scope is wider than that, affecting at least these two targets:

[]

[]

There are hundreds of other failures in the TGP results that we haven't traced to a cause yet, but it's troubling that this CL was implicated twice in the small sampling I did.

[]

*** Original change description ***

Use JavaCompilationArtifacts instead of JavaCompilationArgs

to store Java compilation artifacts. Round-tripping through
JavaCompilationArgs loses dependency information that upstream
compilations use to optimize compile time classpaths.

PiperOrigin-RevId: 159154375