ResourceFilter properly handles special language qualifiers
Two special types of language qualifiers are commonly used:

Serbian can be written in Latin or Cyrillic characters. Serbian in Latin
characters is referred to as sr-Latn (or, with a misplaced region prefix,
sr-rLatn, or with the permutation common in some code for the old verions of
regions supported by Aapt, sr_Latn).

Spanish spoken across Latin America and the Caribbean can be represented by
es-419 (or, using the permutation for old qualifier formats, es_419).

For both of these, transform the resource qualifier into an appropriate BCP-47
representation that can be handled by FolderConfiguration.

Also, add attribute warnings when these qualifiers are used in their specially-handled form (instead of in BCP-47 form).

RELNOTES: none
PiperOrigin-RevId: 160143247