LLVM FDO zipped profiles contain the LLVM raw profile file whose contents are named fdocontrolz_gen_profile.profraw.
LLVM zipped profiles are obtained via fdocontrolz which names the raw
profile file fdocontrol_gen_profile.profraw. Previously, this file was
named to be identical to the zip file but that had to be changed as
users want to retain the ability to rename the zip file.

RELNOTES[NEW]: Change to handle LLVM FDO zipped profile contents correctly.

PiperOrigin-RevId: 160230719