Move the docker tests in src/test/docker
This is a better location and tools/cpp is packaged in embedded_tools that make the test ends-up in embedded_tools, which is requiring network access
to build embedded_tools!

Side-effect: have to correctly mock the bazel workspace file on a couple
of tests.

This is blocking on https://bazel-review.googlesource.com/c/12350/

PiperOrigin-RevId: 160265524